+++ Libplanet.Net.Tests/SwarmTest.Preload.cs
@@ -258,7 +258,7 @@ namespace Libplanet.Net.Tests
         {
             var policy = new BlockPolicy<DumbAction>(
                 new MinerReward(1),
-                getValidators: _ => ConsensusValidators);
+                getValidatorSet: _ => ConsensusValidatorSet);
             var renderer1 = new RecordingActionRenderer<DumbAction>();
             var renderer2 = new RecordingActionRenderer<DumbAction>();
             var chain1 = MakeBlockChain(
@@ -329,7 +329,8 @@ namespace Libplanet.Net.Tests
         [Fact(Timeout = Timeout)]
         public async Task PreloadWithFailedActions()
         {
-            var policy = new BlockPolicy<ThrowException>(getValidators: _ => ConsensusValidators);
+            var policy = new BlockPolicy<ThrowException>(
+                getValidatorSet: _ => ConsensusValidatorSet);
             var fx1 = new MemoryStoreFixture();
             var fx2 = new MemoryStoreFixture();
             var minerChain = MakeBlockChain(policy, fx1.Store, fx1.StateStore);
@@ -398,7 +399,7 @@ namespace Libplanet.Net.Tests
             var fxForNominers = new StoreFixture[2];
             var policy = new BlockPolicy<DumbAction>(
                 new MinerReward(1),
-                getValidators: _ => ConsensusValidators);
+                getValidatorSet: _ => ConsensusValidatorSet);
             fxForNominers[0] = new MemoryStoreFixture(policy.BlockAction);
             fxForNominers[1] = new MemoryStoreFixture(policy.BlockAction);
             var blockChainsForNominers = new[]
@@ -905,7 +906,7 @@ namespace Libplanet.Net.Tests
         {
             var key1 = new PrivateKey();
             var key2 = new PrivateKey();
-            var policy = new BlockPolicy<DumbAction>(getValidators: _ => ConsensusValidators);
+            var policy = new BlockPolicy<DumbAction>(getValidatorSet: _ => ConsensusValidatorSet);
             var genesisContent1 = new BlockContent<DumbAction>(
                 new BlockMetadata(
                     index: 0,
@@ -1004,7 +1005,7 @@ namespace Libplanet.Net.Tests
         {
             var policy = new BlockPolicy<DumbAction>(
                 new MinerReward(1),
-                getValidators: _ => ConsensusValidators);
+                getValidatorSet: _ => ConsensusValidatorSet);
             var fx1 = new MemoryStoreFixture(policy.BlockAction);
             var fx2 = new MemoryStoreFixture(policy.BlockAction);
             var seedChain = MakeBlockChain(policy, fx1.Store, fx1.StateStore);
@@ -1065,7 +1066,7 @@ namespace Libplanet.Net.Tests
             PrivateKey seedKey = new PrivateKey();
             var policy = new BlockPolicy<DumbAction>(
                 new MinerReward(1),
-                getValidators: _ => ConsensusValidators);
+                getValidatorSet: _ => ConsensusValidatorSet);
             var fx1 = new MemoryStoreFixture(policy.BlockAction);
             var fx2 = new MemoryStoreFixture(policy.BlockAction);
             var seedChain = MakeBlockChain(policy, fx1.Store, fx1.StateStore);