+++ Libplanet.Net/Consensus/Context.cs
@@ -82,7 +82,7 @@ namespace Libplanet.Net.Consensus
 
         private readonly BlockChain<T> _blockChain;
         private readonly Codec _codec;
-        private readonly ValidatorSet _validators;
+        private readonly ValidatorSet _validatorSet;
         private readonly Channel<ConsensusMsg> _messageRequests;
         private readonly Channel<System.Action> _mutationRequests;
         private readonly MessageLog _messageLog;
@@ -179,7 +179,7 @@ namespace Libplanet.Net.Consensus
             _preVoteTimeoutFlags = new HashSet<int>();
             _hasTwoThirdsPreVoteFlags = new HashSet<int>();
             _preCommitTimeoutFlags = new HashSet<int>();
-            _validators = validators;
+            _validatorSet = validators;
             _cancellationTokenSource = new CancellationTokenSource();
             ConsensusContext = consensusContext;
             _blockHashCache = new LRUCache<BlockHash, bool>(cacheSize, Math.Max(cacheSize / 64, 8));
@@ -262,7 +262,7 @@ namespace Libplanet.Net.Consensus
             var dict = new Dictionary<string, object>
             {
                 { "node_id", _privateKey.ToAddress().ToString() },
-                { "number_of_validators", _validators.TotalPower },
+                { "number_of_validators", _validatorSet.TotalCount },
                 { "height", Height },
                 { "round", Round },
                 { "step", Step.ToString() },
@@ -435,7 +435,7 @@ namespace Libplanet.Net.Consensus
         private bool HasTwoThirdsPreVote(int round, Func<ConsensusPreVoteMsg, bool> predicate)
         {
             int count = _messageLog.GetPreVotes(round).Count(preVote => predicate(preVote));
-            return count > _validators.TwoThirdsPower;
+            return count > _validatorSet.TwoThirdsCount;
         }
 
         /// <summary>
@@ -449,7 +449,7 @@ namespace Libplanet.Net.Consensus
         private bool HasTwoThirdsPreCommit(int round, Func<ConsensusPreCommitMsg, bool> predicate)
         {
             int count = _messageLog.GetPreCommits(round).Count(preCommit => predicate(preCommit));
-            return count > _validators.TwoThirdsPower;
+            return count > _validatorSet.TwoThirdsCount;
         }
 
         /// <summary>
@@ -462,7 +462,7 @@ namespace Libplanet.Net.Consensus
         /// otherwise <see langword="false"/>.</returns>
         private bool HasOneThirdValidators(int round)
         {
-            return _messageLog.GetValidatorsCount(round) > _validators.OneThirdPower;
+            return _messageLog.GetValidatorsCount(round) > _validatorSet.OneThirdCount;
         }
     }
 }