+++ Libplanet/Blockchain/Policies/NullBlockPolicy.cs
@@ -5,6 +5,7 @@ using System.Collections.Immutable;
 using Libplanet.Action;
 using Libplanet.Assets;
 using Libplanet.Blocks;
+using Libplanet.Consensus;
 using Libplanet.Crypto;
 using Libplanet.Tx;
 
@@ -15,16 +16,16 @@ namespace Libplanet.Blockchain.Policies
     {
         private readonly BlockPolicyViolationException _exceptionToThrow;
         private readonly long _difficulty;
-        private readonly Func<long, IEnumerable<PublicKey>> _getValidators;
+        private readonly Func<long, ValidatorSet> _getValidatorSet;
 
         public NullBlockPolicy(
             BlockPolicyViolationException exceptionToThrow = null,
             long difficulty = 1,
-            Func<long, IEnumerable<PublicKey>> getValidators = null)
+            Func<long, ValidatorSet> getValidatorSet = null)
         {
             _exceptionToThrow = exceptionToThrow;
             _difficulty = difficulty;
-            _getValidators = getValidators ?? (_ => ImmutableList<PublicKey>.Empty);
+            _getValidatorSet = getValidatorSet ?? (_ => new ValidatorSet(new List<PublicKey>()));
         }
 
         public ISet<Address> BlockedMiners { get; } = new HashSet<Address>();
@@ -62,6 +63,6 @@ namespace Libplanet.Blockchain.Policies
         public int GetMaxTransactionsPerSignerPerBlock(long index) =>
             GetMaxTransactionsPerBlock(index);
 
-        public IEnumerable<PublicKey> GetValidators(long index) => _getValidators(index);
+        public ValidatorSet GetValidatorSet(long index) => _getValidatorSet(index);
     }
 }